Add MarkPending to the repository
Note that the mock is just returning 'nil' but presumably that will
change when using MarkPending in the actual function.

Note that the migration is so complicated because you can't remove an
enum from an existing type.  You can add it, but not in a transaction so
it doesn't work in a migration.

Add an index querying pending